* Remove long values * Add name / logo url to developers * Add name / logo url fallback for tenants * Add material design colors to developers and tenants

+++ HCore-Database/ElasticSearch/IElasticSearchDbContext.cs
@@ -7,6 +7,6 @@ namespace HCore.Database.ElasticSearch
         string[] IndexNames { get; }
 
         CreateIndexDescriptor GetCreateIndexDescriptor(IElasticSearchClient elasticSearchClient, string indexName);
-        long GetIndexVersion(string indexName);
+        int GetIndexVersion(string indexName);
     }
 }
